requirements.txt
asgiref==3.10.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==42.0.8
dj-database-url==3.0.1
Django==5.0.2
django-compat==1.0.15
django-cors-headers==4.7.0
django-filter==25.1
django-rest-auth==0.9.5
djangorestframework==3.16.1
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-storage==2.15.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==10.4.0
proto-plus==1.26.1
protobuf==6.33.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
python-dotenv
